+++ js/salesExecutiveApi.js
@@ -11,8 +11,11 @@ let deleteButtons;
 const sidebarButtonToggle = document.getElementById("sidebar-toggle");
 
 sidebarButtonToggle.addEventListener("click", () => {
-  const sidebar = document.querySelector("[data-sidebar-style='overlay'] .dlabnav");
-  if(!sidebar.style.left || sidebar.style.left === "-100%") sidebar.style.left = 0;
+  const sidebar = document.querySelector(
+    "[data-sidebar-style='overlay'] .dlabnav"
+  );
+  if (!sidebar.style.left || sidebar.style.left === "-100%")
+    sidebar.style.left = 0;
   else sidebar.style.left = "-100%";
 });
 
@@ -40,7 +43,9 @@ const getSalesExecs = (pagesize = pageSize, page = 1, search = "") => {
       salesExecutiveTableInfo.innerText = `Showing ${
         (json.currentpage - 1) * pagesize + 1
       } to ${
-        pagesize < pagesize * page + salesExecs.length ? pagesize * page : pagesize
+        pagesize < pagesize * page + salesExecs.length
+          ? pagesize * page
+          : pagesize
       } of ${json.totalpages * pagesize} entries`;
 
       for (let i = 0; i < salesExecs.length; i++) {
@@ -88,22 +93,24 @@ const getSalesExecs = (pagesize = pageSize, page = 1, search = "") => {
       deleteButtons.forEach((button) =>
         button.addEventListener("click", () => {
           const _id = button.getAttribute("data-id");
-          fetch(
-            `https://pg-app-backend.herokuapp.com/api/superadmin/salesexecutive/${_id}`,
-            {
-              method: "DELETE",
-              headers: {
-                Authorization: `Bearer ${
-                  JSON.parse(localStorage.getItem("user")).token
-                }`,
-              },
-            }
-          )
-            .then((res) => res.json())
-            .then((data) => {
-              if(data.error) return window.alert(data.error);
-              window.location.reload();
-            });
+          if (window.confirm("Are you sure?")) {
+            fetch(
+              `https://pg-app-backend.herokuapp.com/api/superadmin/salesexecutive/${_id}`,
+              {
+                method: "DELETE",
+                headers: {
+                  Authorization: `Bearer ${
+                    JSON.parse(localStorage.getItem("user")).token
+                  }`,
+                },
+              }
+            )
+              .then((res) => res.json())
+              .then((data) => {
+                if (data.error) return window.alert(data.error);
+                window.location.reload();
+              });
+          }
         })
       );
     });
@@ -147,7 +154,6 @@ searchBar.addEventListener("blur", () => {
 
 getSalesExecs();
 loadPaginationEventListeners();
-
 
 // post new sales exec
 const newSalesExecNameInput = document.getElementById(
