+++ js/userDatatableApi.js
@@ -12,8 +12,11 @@ let deleteButtons;
 const sidebarButtonToggle = document.getElementById("sidebar-toggle");
 
 sidebarButtonToggle.addEventListener("click", () => {
-  const sidebar = document.querySelector("[data-sidebar-style='overlay'] .dlabnav");
-  if(!sidebar.style.left || sidebar.style.left === "-100%") sidebar.style.left = 0;
+  const sidebar = document.querySelector(
+    "[data-sidebar-style='overlay'] .dlabnav"
+  );
+  if (!sidebar.style.left || sidebar.style.left === "-100%")
+    sidebar.style.left = 0;
   else sidebar.style.left = "-100%";
 });
 
@@ -94,25 +97,26 @@ const getUsers = (pagesize = pageSize, page = 1, search = "") => {
       deleteButtons.forEach((button) =>
         button.addEventListener("click", () => {
           const _id = button.getAttribute("data-id");
-          fetch(
-            `https://pg-app-backend.herokuapp.com/api/superadmin/user/${_id}`,
-            {
-              method: "DELETE",
-              headers: {
-                Authorization: `Bearer ${
-                  JSON.parse(localStorage.getItem("user")).token
-                }`,
-              },
-            }
-          )
-            .then((res) => res.json())
-            .then((data) => {
-              if(data.error) return window.alert(data.error);
-              window.location.reload();
-            });
+          if (window.confirm("Are you sure?")) {
+            fetch(
+              `https://pg-app-backend.herokuapp.com/api/superadmin/user/${_id}`,
+              {
+                method: "DELETE",
+                headers: {
+                  Authorization: `Bearer ${
+                    JSON.parse(localStorage.getItem("user")).token
+                  }`,
+                },
+              }
+            )
+              .then((res) => res.json())
+              .then((data) => {
+                if (data.error) return window.alert(data.error);
+                window.location.reload();
+              });
+          }
         })
       );
-
     });
 };
